make

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: falves-e <falves-e@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/16 17:23:43 by falves-e          #+#    #+#             */
-/*   Updated: 2026/04/20 13:20:57 by falves-e         ###   ########.fr       */
+/*   Updated: 2026/04/21 16:29:46 by falves-e         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int			i;
+	size_t		i;
 	const char	*str1;
 	const char	*str2;
